@@ -9,39 +9,38 @@
         "ext-gd": "*",
         "ext-json": "*",
         "ext-mbstring": "*",
-        "algolia/algoliasearch-client-php": "^2.7",
-        "algolia/scout-extended": "^1.15",
-        "babenkoivan/elastic-migrations": "^1.4",
-        "babenkoivan/elastic-scout-driver": "^1.3",
-        "babenkoivan/elastic-scout-driver-plus": "^2.0",
-        "cloudcreativity/laravel-json-api": "^3.2",
-        "cviebrock/eloquent-sluggable": "^8.0",
+        "algolia/algoliasearch-client-php": "^3.2",
+        "algolia/scout-extended": "^2.0",
+        "babenkoivan/elastic-migrations": "^2.0",
+        "babenkoivan/elastic-scout-driver": "^2.0",
+        "babenkoivan/elastic-scout-driver-plus": "^3.3",
+        "cloudcreativity/laravel-json-api": "^4.0",
+        "cviebrock/eloquent-sluggable": "^9.0",
         "doctrine/dbal": "^3.0",
-        "fideloper/proxy": "^4.4",
         "fruitcake/laravel-cors": "^2.0",
         "guzzlehttp/guzzle": "^7.0.1",
-        "laravel/framework": "^8.12",
-        "laravel/scout": "^8.6",
-        "laravel/tinker": "^2.5",
-        "league/flysystem-aws-s3-v3": "~1.0",
+        "laravel/framework": "^9.0",
+        "laravel/scout": "^9.0",
+        "laravel/tinker": "^2.7",
+        "league/flysystem-aws-s3-v3": "^3.0",
         "phospr/fraction": "^1.2",
         "spatie/laravel-csp": "^2.6",
-        "spatie/laravel-medialibrary": "^9.0.0",
-        "spatie/laravel-tags": "^3.0"
+        "spatie/laravel-medialibrary": "^10.0",
+        "spatie/laravel-tags": "^4.0"
     },
     "require-dev": {
         "barryvdh/laravel-ide-helper": "^2.9",
         "brianium/paratest": "^6.2",
-        "cloudcreativity/json-api-testing": "^3.2",
-        "facade/ignition": "^2.5",
+        "cloudcreativity/json-api-testing": "^4.0",
         "fakerphp/faker": "^1.9.1",
         "laravel/breeze": "^1.0",
         "laravel/sail": "^1.10",
         "mockery/mockery": "^1.4.2",
-        "nunomaduro/collision": "^5.0",
-        "nunomaduro/larastan": "^0.6.13",
+        "nunomaduro/collision": "^6.1",
+        "nunomaduro/larastan": "^2.0",
         "php-coveralls/php-coveralls": "^2.4",
-        "phpunit/phpunit": "^9.3.3"
+        "phpunit/phpunit": "^9.3.3",
+        "spatie/laravel-ignition": "^1.0"
     },
     "config": {
         "optimize-autoloader": true,
