@@ -9,12 +9,11 @@ use App\Models\Traits\Sluggable;
 use App\Models\Traits\Taggable;
 use App\Support\Number;
 use App\Support\Nutrients;
-use ElasticScoutDriverPlus\QueryDsl;
+use ElasticScoutDriverPlus\Searchable;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Support\Collection;
-use Laravel\Scout\Searchable;
 use Spatie\Image\Manipulations;
 use Spatie\MediaLibrary\HasMedia;
 use Spatie\MediaLibrary\InteractsWithMedia;
@@ -91,7 +90,6 @@ final class Recipe extends Model implements HasMedia
     use Ingredient;
     use InteractsWithMedia;
     use Journalable;
-    use QueryDsl;
     use Searchable;
     use Sluggable;
     use Taggable;
